flake.{lock,nix}, devshell.toml: remove
This can be removed - the only thing that we need is a go compiler (to
build and run unit tests).

gopls is editor tooling, awscli and google-cloud-sdk is unused.
minio/mc were used to provide a s3-compatible server for integration
tests, but got removed in the previous commit, too

Currently, this is simply a go libary, and the only requirement is a go
compiler. We bring in different versions of it through CI, cause we want
to test compatibility with (currently multiple) versions of it.