feat(pkg/nar): rewrite
This rewrites the NAR parser to a recursive version, running in its own
goroutine, which seems to be way easier to understand than the previous
iterative version.

We use a channel to communicate with the outside world, so the iterative
interface can be preserved while internally having recursion instead of
manually managing stacks.

Essentially, the structure we parse looks like this:

<Node>
 - ( type regular executable contents … )
 - ( type symlink target foo )
 - ( type directory
     entry ( name foo1 node <Node> )
     ...
   )

This also renames EntryType to NodeType.

As can be seen from the grammar, a Node has a type (regular, symlink or
directory).

In case the Node is a directory, it has a list of entries.
Each Entry has a name and a Node, but no type.
An Entry can be seen as a map from Name to Node, which is now reflected
in pkgs/nar/ls as well.

Co-Authored-By: Félix Baylac-Jacqué <felix@alternativebit.fr>