feat(pkg/nar/header.go): add Header.Validate() function
This does provide some consistency checking of the header structure,
such as checking for valid paths and inconsistent fields, and returns an
error if it fails validation.

It's used in the NAR Writer, so we return meaningful errors when trying
to write an invalid header.

We don't add it into the NAR Reader, as there's no way for it to emit
these broken headers.