pkg/derivation: align Derivation struct with json format
A lot of the struct structure has been designed to work with the parser
generator we used.

Since #55 we have our own parser, so we can be a bit more flexible with
how the structure of the Derivation struct looks like.

Use the same structure that `nix show-derivation` also exposes.

Outputs, InputDerivations and Env are maps, not lists with odd order
requirements. The Parser takes care of ensuring the order was correct
when parsing it in, and the encoder outputs them in a sorted fashion.

This makes the Validate() method much smaller, as there's simply less
invalid states someone manually populating the struct could produce.