pkg/derivation: make Derivation store context-aware
Propagating contexts allows cancelling fetching operations. While not so
interesting for local stores, they make a lot of sense for HTTP stores.

This does bleed a bit through to
Derivation.{GetSubstitutionHash,OutputPaths} and
Derivation.writeDerivation, as we need to pass the context back to when
we query the store again.