pkg/derivation: Reduce allocations in writeDerivation
Instead of allocating a bunch of arrays in which we encode the output
we can write the output directly to the writer.

This sacrifices readability a bit and I'm not too happy about the
excessive error handling required but I think it's worth it given
the quite drastic performance improvement.

Using the following program for testing:
``` go
package main

import (
	"fmt"
	"os"
	"testing"
	"time"

	"github.com/nix-community/go-nix/pkg/derivation"
)

type nullWriter struct{}

func (n *nullWriter) Write(b []byte) (int, error) {
	return len(b), nil
}

func main() {

	drvPath := "/nix/store/lfb109iifgph2xa94aps13n2w8bpbqsq-texlive-combined-full-2021-final.drv"

	n := 50

	derivationFile, err := os.Open(drvPath)
	if err != nil {
		panic(err)
	}

	drv, err := derivation.ReadDerivation(derivationFile)
	if err != nil {
		panic(err)
	}

	writer := &nullWriter{}

	{
		start := time.Now()

		for i := 0; i < n; i++ {
			err = drv.WriteDerivation(writer)
			if err != nil {
				panic(err)
			}
		}

		fmt.Println(time.Since(start) / time.Duration(n))
	}

	fmt.Println("Allocs:", int(testing.AllocsPerRun(n, func() {
		err = drv.WriteDerivation(writer)
		if err != nil {
			panic(err)
		}
	})))

}
```

Yields the following performance numbers:
- Before
```
3.024143ms
Allocs: 15508
```

- After
```
1.279567ms
Allocs: 86
```

Note that I picked `texlive.combined.scheme-full` which is pretty much
the largest derivation file imaginable that is realistic and found out in
the wild.
The number of allocations is static and doesn't change when writing
another smaller derivation.

!