pkg/exp/store: replace BlobHasher with blobWriter
This can be used to build a streaming interface for stores to write
blobs to, that also takes care of hashing.

We actually don't need any of the Reset() functionality, as we already
manage this for the underlying hash before putting things back in the
pool sync.Pool, and can ask the hash for Size() and BlockSize()
directly.