pkg/exp/store/blobstore: filestore: fix writing the same blob while reading it
Windows doesn't like writing over already opened files. If we end up
writing a blob into the FileStore that already exist, and is currently
being read, the file is opened, and closing the writer from WriteBlob
fails, as it can't move the temporary file to the new location.

Do a stat to figure out if dstPath already exist, and simply discard the
temporary if we already have that blob.