pkg/exp/store: replace treestore.DirEntryPath with treestore.Entry
This is a simple struct containing an fs.DirEntry, not an interface also
implementing it.

This allows us keeping around the original fs.DirEntry instances,
instead of trying to copy their attributes around.

We still keep a mockDirEntry, mockDirFileInfo, but for testing purposes
only. We put them in the fixture package.

With the change to that, we can now also easily test
importer.FromFilesystemFilter, without treestore.BuildTree in the loop,
so move the `[]treestore.Entry` sclices produced into the fixture
package as well.