Add support for recursive search
The Nix, Lix & Snix (twix) codebase both support a recursive call
for the lookup of the input derivation replacement paths.

Augment hashes.go to include the ability to recursively lookup
derivations and calculate their paths.

A demonstration of this is done in the test but it can also be done
by looking up derivations on the /nix/store filesystem for instance.